Separated f32 and f64 versions of frexp, ilogb, and ldexp. Also optimized frexpl, ilogbl, and ldexpl
