+++ src/libc/frexpl.src
@@ -1,0 +1,19 @@
+	assume	adl=1
+
+	section	.text
+
+	public	_frexpl
+
+; int frexpl(long double, int*)
+_frexpl:
+	ld	hl, 10
+	add	hl, sp
+	bit	7, (hl)	; signbit(x)
+	jp	z, __frexpl_c	; positive input
+	; negative input
+	res	7, (hl)	; x = fabsl(x)
+	call	__frexpl_c
+	set 7, b	; copysignl(x, -1.0L)
+	ret
+
+	extern	__frexpl_c
